[EP-2124] Changes after code review
