+++ packages/earth-map/src/components/header/index.ts
@@ -1,9 +1,10 @@
 import OrgSwitcher from './component';
 import { connect } from 'react-redux';
 import { setUserGroup } from 'modules/user/actions';
-import { resetPlacesFeatured, setPlacesSearch } from 'modules/places/actions';
+import { resetPlacesFeatured, setPlacesSearch, resetPlace} from 'modules/places/actions';
 import { resetLayerCache } from 'modules/layers/actions';
 import { resetMap } from 'modules/map/actions';
+import { setIndexesSelected } from 'modules/indexes/actions';
 
 export default connect(
   (state: any) => ({
@@ -13,8 +14,10 @@ export default connect(
   {
     setUserGroup,
     setPlacesSearch,
+    resetPlace,
     resetPlacesFeatured,
     resetLayerCache,
-    resetMap
+    resetMap,
+    setIndexesSelected,
   }
 )(OrgSwitcher);
