+++ packages/earth-map/src/components/header/styles.scss
@@ -19,6 +19,7 @@
     text-overflow: ellipsis;
   }
 }
+
 .ng-org-badge {
   background: $marapp-gray-1;
   border-radius: 50%;
@@ -30,4 +31,21 @@
   top: 0;
   text-align: center;
   width: 20px;
+}
+
+.ng-ep-kicker {
+  margin-left: 10px;
+  margin-right: 15px;
+  position: relative;
+  &:before {
+    content: '';
+    background: $marapp-gray-0;
+    display: block;
+    height: 10px;
+    position: absolute;
+    transform: translateY(-50%);
+    top: 0;
+    bottom: 0;
+    width: 2px;
+  }
 }
