+++ packages/earth-map/src/components/sidebar/sidebar-toggle/styles.scss
