+++ packages/earth-map/src/components/sidebar/styles.scss
@@ -25,20 +25,3 @@
     width: 500px;
   }
 }
-
-.ng-ep-kicker {
-  margin-left: 10px;
-  margin-right: 15px;
-  position: relative;
-  &:before {
-    content: '';
-    background: $marapp-gray-0;
-    display: block;
-    height: 10px;
-    position: absolute;
-    transform: translateY(-50%);
-    top: 0;
-    bottom: 0;
-    width: 2px;
-  }
-}
